Move IndexSequence to _support.indexing
