+++ shark_turbine/kernel/ops/wave_ops.py
@@ -18,8 +18,8 @@ from typing import (
 import torch.fx as fx
 
 if TYPE_CHECKING:
-    from ..lang.wave_types import Memory, Register, IndexMapping
-from .._support.indexing import IndexExpr, IndexSymbol
+    from ..lang.wave_types import Memory, Register
+from .._support.indexing import IndexExpr, IndexSymbol, IndexSequence
 from .._support.dtype import DataType
 from .._support.regions import RegionGraph
 from .base import OpDispatcher
